Refactor Metagraph component to integrate MetagraphEffectsCoordinator and replace MetagraphFocusMode with MetagraphClustersUI
